rename fabric directory to memfabric
in addition:
 - make use of the new logging functions instead of DPRINTF
