+++ projects/rocshmem/src/CMakeLists.txt
@@ -80,4 +80,4 @@ add_subdirectory(containers)
 add_subdirectory(memory)
 add_subdirectory(sync)
 add_subdirectory(bootstrap)
-add_subdirectory(fabric)
+add_subdirectory(memfabric)
